Disable p10k instant prompt
pretty sure that the instant prompt is messing w/ the gpg env var
causing errors when trying to sign commits. I believe the instant prompt
is still sometimes active when the GPG_TTY=$(tty) line is executed
and the instant prompt is not a tty environment, so we cannot sign
commits. Hopefully removing this solves the problem. If not, will re-add
and debug this problem elsewhere.